move cow api to its own folder
