+++ src/utils/price.ts
@@ -1,5 +1,5 @@
 import { GetQuoteResponse } from '@gnosis.pm/gp-v2-contracts'
-import { OrderMetaData } from '../api/types'
+import { OrderMetaData } from '../api/cow/types'
 import { SupportedChainId as ChainId } from '../constants/chains'
 
 export interface QuoteParams {
